+++ CHANGELOG.md
@@ -40,6 +40,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Improvements
 
 * [#709](https://github.com/babylonlabs-io/finality-provider/pull/709) chore: query indexed block
+* [#711](https://github.com/babylonlabs-io/finality-provider/pull/711) chore: improve proof batch building
+
 
 ## v2.0.0-rc.4