+++ CHANGELOG.md
@@ -41,7 +41,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 * [#709](https://github.com/babylonlabs-io/finality-provider/pull/709) chore: query indexed block
 * [#711](https://github.com/babylonlabs-io/finality-provider/pull/711) chore: improve proof batch building
-
+* [#715](https://github.com/babylonlabs-io/finality-provider/pull/715) feat: batch eots sign
 
 ## v2.0.0-rc.4