+++ CHANGELOG.md
@@ -37,6 +37,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+## v2.0.0-rc.4
+
 ### Improvements
 
 * [#709](https://github.com/babylonlabs-io/finality-provider/pull/709) chore: query indexed block